chore: replaced any and as assertions with types
