+++ src/chart_types/partition_chart/layout/types/config_types.ts
@@ -129,44 +129,3 @@ export interface Config extends StaticConfig {
     keyframes: Array<AnimKeyframe>;
   };
 }
-
-// switching to `io-ts` style, generic way of combining static and runtime type info - 1st step
-class Type<A> {
-  dflt: A;
-
-  reconfigurable: boolean | string;
-
-  documentation = 'string';
-
-  constructor(dflt: A, reconfigurable: boolean | string, documentation: string) {
-    this.dflt = dflt;
-    this.reconfigurable = reconfigurable;
-    this.documentation = documentation;
-  }
-}
-
-export class Numeric extends Type<number> {
-  min: number;
-
-  max: number;
-
-  type = 'number';
-
-  constructor({
-    dflt,
-    min,
-    max,
-    reconfigurable,
-    documentation,
-  }: {
-    dflt: number;
-    min: number;
-    max: number;
-    reconfigurable: boolean | string;
-    documentation: string;
-  }) {
-    super(dflt, reconfigurable, documentation);
-    this.min = min;
-    this.max = max;
-  }
-}
